+++ Scripts/Enemy.gd
@@ -236,7 +236,7 @@ func _death():
 func leave_behind_item():
     var leave_percentage = constants.ENEMY_LEAVE_BEHIND_ITEM_PERCENTAGE + (get_parent().get_node('Player').upgrades['LOOT LOVER'][0] * 10)
     
-    if randi_range(1,100) < leave_percentage:
+    if randi_range(1,100) <= leave_percentage:
         var item = get_parent().item_scene.instantiate()
         get_parent().add_child(item)
         item.spawn_random()