+++ Scripts/MainMenu.gd
@@ -5,6 +5,7 @@ signal exit_game_pressed
 signal credits_pressed
 signal cheats_pressed
 signal game_mode_pressed
+signal statistics_pressed
 
 # Called when the node enters the scene tree for the first time.
 func _ready():
@@ -31,3 +32,6 @@ func _on_credits_pressed():
     
 func _on_game_mode_pressed():
     emit_signal('game_mode_pressed')
+
+func _on_statistics_pressed():
+    emit_signal('statistics_pressed')